backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.3
httpx==0.27.2
python-multipart==0.0.9
structlog==24.4.0
orjson==3.10.7
boto3==1.35.25
Pillow==10.4.0
python-barcode==0.15.1
weasyprint==62.3
openai==1.50.2
pgvector==0.3.2
tenacity==9.0.0
pandas==2.2.2
openpyxl==3.1.5
redis==5.1.1
celery==5.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
respx==0.21.1
aiosqlite==0.20.0
